requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
ultralytics==8.1.0
opencv-python==4.8.1.78
torch==2.1.1
torchvision==0.16.1
torchreid==1.4.0
numpy==1.24.3
Pillow==10.1.0
supabase==2.3.0
python-telegram-bot==20.7
twilio==8.10.0
scikit-learn==1.3.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6

